connect course and auth to backend

+++ src/course/DeleteCourse.js
@@ -1,31 +1,50 @@
 import { useMemo, useState } from "react";
-import axios from "axios";
 
 import { Dialog, Description } from "@headlessui/react";
-import { FallingLines } from "react-loader-spinner";
-import { BiError } from "react-icons/bi";
+import { toast } from "react-toastify";
+import { FaCircleNotch } from "react-icons/fa";
+import axios from "axios";
+import { useLocation, useNavigate } from "react-router-dom";
 
 const MAX_TITLE_CHARS = 25;
 
-export function DeleteCourseBtn({courseId, courseTitle}) {
+export function DeleteCourseBtn({courseId, courseTitle, className=""}) {
     const [isOpen, setIsOpen] = useState(false);
     const [retypeTitle, setRetypeTitle] = useState("");
     const [isLoading, setIsLoading] = useState(false);
-    const [errors, setErrors] = useState({});
     
+    const { pathname } = useLocation();
+    const navigate = useNavigate();
+
     const TITLE_RETYPE_TEST = useMemo(
         () => courseTitle.substring(0, MAX_TITLE_CHARS).trim(), 
         [courseTitle]
     );
 
     async function handleDelete() {
+        if (isLoading) return;
+        setIsLoading(true);
+
         const beUrl = process.env.REACT_APP_COURSE_BE + 
                         "/course/delete?id=" + courseId;
         
-        setIsLoading(true);
-        closeMainDialog();
         try {
             await axios.delete(beUrl);
+            
+            closeMainDialog();
+            
+            navigate(pathname, {replace: true});
+
+            const message = <>
+                <p>Course <strong>{TITLE_RETYPE_TEST}</strong> has been deleted.</p>
+                <p>Please, <strong>REFRESH</strong> this page to see the update</p>
+            </>
+
+            toast(message, {
+                type: "info",
+                autoClose: 5000,
+            });
+
         } catch(error) {
             let message = "";
             if (error.response) {
@@ -33,9 +52,16 @@ export function DeleteCourseBtn({courseId, courseTitle}) {
             } else {
                 message = "Network Error. Please check again your connection";
             }
-            setErrors({
-                message: message
-            })
+            toast(
+                message, 
+                {
+                    type: "error",
+                    autoClose: false,
+                    hideProgressBar: true
+                }
+            );
+            throw error;
+
         } finally {
             setIsLoading(false);
         }
@@ -54,37 +80,15 @@ export function DeleteCourseBtn({courseId, courseTitle}) {
     return (
         <>
             <button 
-                className="text-white backdrop-blur-md rounded-md 
-                            bg-red-600 p-2 font-bold hover:bg-red-400"
+                className={`text-white backdrop-blur-md rounded-md 
+                            bg-red-500 p-2 font-bold hover:bg-red-800 ${className}`}
                 onClick={openMainDialog}>
                 Delete
             </button>
-            <Dialog open={isLoading} onClose={() => setIsLoading(false)} className="relative z-60">
-                <div className="fixed inset-0 flex w-screen h-screen items-center justify-center p-4 backdrop-blur-sm">
-                    <FallingLines color="gray" width="150"/>
-                </div>
-            </Dialog>
             
-            <Dialog open={errors.message !== undefined} onClose={() => setErrors({})} className="relative z-60">
-                <div className="fixed inset-0 flex w-screen h-screen items-center justify-center p-4 backdrop-blur-sm">
-                    <div className="p-4 w-fit bg-white rounded-md flex shadow-md shadow-gray/80">
-                        <BiError className="self-center mx-auto size-[4rem]" color="red"/>
-                        <div className="relative pl-2 flex flex-col gap-2">
-                            <p className="text-[1rem] md:text-[1.8rem] font-semibold">{errors.message}</p>
-                            <button 
-                                onClick={() => setErrors({})}
-                                className="self-end w-fit shadow-inner shadow-white/10 py-1.5 px-3 rounded-md
-                                             font-semibold text-white bg-gray-400 hover:bg-gray-700">
-                                Close
-                            </button>
-                        </div>
-                    </div>
-                </div>
-            </Dialog>
-
             <Dialog open={isOpen} onClose={closeMainDialog} className="relative z-50">
                 <div className="fixed inset-0 flex w-screen items-center justify-center p-4 backdrop-blur-sm">
-                    <div className="max-w-lg space-y-4 border bg-slate-100/95 p-12">
+                    <div className="max-w-lg space-y-4 border bg-slate-100 p-12">
                         <div className="font-bold">Delete Course</div>
                         <Description className="line-clamp-2">
                             This will permanently delete course <strong>{courseTitle}</strong>
@@ -97,16 +101,23 @@ export function DeleteCourseBtn({courseId, courseTitle}) {
                         <input className="w-full rounded-md text-sm h-fit shadow-sm" onChange={(e) => setRetypeTitle(e.target.value)}/>
                         <div className="flex gap-4 justify-end">
                             <button 
-                                className="p-2 bg-gray-500 text-white font-semibold rounded-sm" 
+                                className="p-2 bg-gray-500 hover:bg-gray-600 text-white font-semibold rounded-sm" 
                                 onClick={closeMainDialog}>
                                     Cancel
                             </button>
                             <button 
-                                className="p-2 bg-red-500 text-white font-semibold rounded-sm
-                                            disabled:bg-red-300 hover:bg-red-500/60"
-                                onClick={handleDelete}
-                                disabled={retypeTitle !== TITLE_RETYPE_TEST}>
-                                    Delete
+                                    className="inline-flex gap-2 p-2 bg-red-500 
+                                                text-white font-semibold rounded-sm
+                                                disabled:bg-red-300 hover:bg-red-800"
+                                    onClick={handleDelete}
+                                    disabled={retypeTitle !== TITLE_RETYPE_TEST}>                                    
+                                
+                                { 
+                                    isLoading &&
+                                        <FaCircleNotch className="animate-spin h-full w-auto min-h-5" color="white"/> 
+                                } 
+                                Delete
+
                             </button>
                         </div>
                     </div>
